# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients/{patientId}/insurance": {
      "post": {
        "operationId": "PatientInsuranceController_create",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientInsuranceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientInsuranceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": "Patient or follow-up not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record patient insurance",
        "tags": [
          "Patient insurance"
        ]
      },
      "get": {
        "operationId": "PatientInsuranceController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientInsuranceResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List patient insurance history",
        "tags": [
          "Patient insurance"
        ]
      }
    },
    "/patients/{patientId}/medical-appointments": {
      "post": {
        "operationId": "PatientMedicalAppointmentsController_create",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientMedicalAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientMedicalAppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": "Patient or follow-up not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a patient medical appointment",
        "tags": [
          "Patient medical appointments"
        ]
      },
      "get": {
        "operationId": "PatientMedicalAppointmentsController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientMedicalAppointmentResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a patient medical appointment history",
        "tags": [
          "Patient medical appointments"
        ]
      }
    },
    "/patients/{patientId}/sis-affiliations": {
      "post": {
        "operationId": "PatientSisAffiliationController_create",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientSisAffiliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSisAffiliationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": "Patient or follow-up not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a patient SIS affiliation",
        "tags": [
          "Patient SIS affiliations"
        ]
      },
      "get": {
        "operationId": "PatientSisAffiliationController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientSisAffiliationResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List patient SIS affiliation history",
        "tags": [
          "Patient SIS affiliations"
        ]
      }
    },
    "/patients/{patientId}/treatments": {
      "post": {
        "operationId": "PatientTreatmentsController_create",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientTreatmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientTreatmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": "Patient, follow-up, or diagnosis not found"
          },
          "409": {
            "description": "Diagnosis does not belong to the patient"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a patient treatment",
        "tags": [
          "Patient treatments"
        ]
      },
      "get": {
        "operationId": "PatientTreatmentsController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientTreatmentResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a patient treatment history",
        "tags": [
          "Patient treatments"
        ]
      }
    },
...
  },
  "info": {
    "title": "FPC Backend API",
    "description": "FPC backend API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePatientInsuranceDto": {
        "type": "object",
        "properties": {
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "insuranceType": {
            "type": "string",
            "enum": [
              "SIS",
              "ESSALUD",
              "EPS",
              "FUERZAS_ARMADAS",
              "SALUDPOL",
              "NONE"
            ]
          },
          "epsProvider": {
            "type": "string",
            "enum": [
              "RIMAC",
              "PACIFICO",
              "MAPFRE",
              "SANITAS",
              "LA_POSITIVA",
              "ONCOSALUD",
              "OTHER"
            ]
          },
          "changeReason": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "followUpId",
          "insuranceType"
        ]
      },
      "PatientInsuranceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "insuranceType": {
            "enum": [
              "SIS",
              "ESSALUD",
              "EPS",
              "FUERZAS_ARMADAS",
              "SALUDPOL",
              "NONE"
            ],
            "type": "string"
          },
          "epsProvider": {
            "nullable": true,
            "enum": [
              "RIMAC",
              "PACIFICO",
              "MAPFRE",
              "SANITAS",
              "LA_POSITIVA",
              "ONCOSALUD",
              "OTHER"
            ],
            "type": "string"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "changeReason": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "followUpId",
          "insuranceType",
          "epsProvider",
          "isCurrent",
          "changeReason",
          "startDate",
          "endDate",
          "createdAt"
        ]
      },
      "CreatePatientMedicalAppointmentDto": {
        "type": "object",
        "properties": {
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "healthCenterId": {
            "type": "string",
            "format": "uuid"
          },
          "specialty": {
            "type": "string"
          },
          "appointmentDate": {
            "type": "string"
          },
          "nextAppointmentDate": {
            "type": "string"
          },
          "hasReferralSheet": {
            "type": "boolean"
          },
          "referredTo": {
            "type": "string"
          },
          "difficulties": {
            "type": "string"
          },
          "isFirstConsultation": {
            "type": "boolean"
          },
          "changeReason": {
            "type": "string"
          }
        },
        "required": [
          "followUpId",
          "specialty"
        ]
      },
      "PatientMedicalAppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "healthCenterId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "specialty": {
            "type": "string"
          },
          "appointmentDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "nextAppointmentDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "hasReferralSheet": {
            "type": "boolean"
          },
          "referredTo": {
            "type": "string",
            "nullable": true
          },
          "difficulties": {
            "type": "string",
            "nullable": true
          },
          "isFirstConsultation": {
            "type": "boolean"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "changeReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "followUpId",
          "healthCenterId",
          "specialty",
          "appointmentDate",
          "nextAppointmentDate",
          "hasReferralSheet",
          "referredTo",
          "difficulties",
          "isFirstConsultation",
          "isCurrent",
          "changeReason",
          "createdAt"
        ]
      },
      "CreatePatientSisAffiliationDto": {
        "type": "object",
        "properties": {
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "canAffiliate": {
            "type": "boolean"
          },
          "expectedDate": {
            "type": "string"
          },
          "cantAffiliateReason": {
            "type": "string"
          },
          "affiliatedAt": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          }
        },
        "required": [
          "followUpId",
          "canAffiliate"
        ]
      },
      "PatientSisAffiliationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "canAffiliate": {
            "type": "boolean"
          },
          "expectedDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "cantAffiliateReason": {
            "type": "string",
            "nullable": true
          },
          "affiliatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "followUpId",
          "canAffiliate",
          "expectedDate",
          "cantAffiliateReason",
          "affiliatedAt",
          "comments",
          "createdAt"
        ]
      },
      "CreatePatientTreatmentDto": {
        "type": "object",
        "properties": {
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "diagnosisId": {
            "type": "string",
            "format": "uuid"
          },
          "treatmentType": {
            "type": "string"
          },
          "treatmentFrequency": {
            "type": "string"
          },
          "healthCenterId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "changeReason": {
            "type": "string"
          },
          "notReceivingReason": {
            "type": "string"
          },
          "treatmentSituation": {
            "type": "string"
          }
        },
        "required": [
          "followUpId",
          "diagnosisId",
          "treatmentType"
        ]
      },
      "PatientTreatmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "diagnosisId": {
            "type": "string",
            "format": "uuid"
          },
          "treatmentType": {
            "type": "string"
          },
          "treatmentFrequency": {
            "type": "string",
            "nullable": true
          },
          "healthCenterId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "changeReason": {
            "type": "string",
            "nullable": true
          },
          "notReceivingReason": {
            "type": "string",
            "nullable": true
          },
          "treatmentSituation": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "followUpId",
          "diagnosisId",
          "treatmentType",
          "treatmentFrequency",
          "healthCenterId",
          "startDate",
          "endDate",
          "isCurrent",
          "changeReason",
          "notReceivingReason",
          "treatmentSituation",
          "createdAt"
        ]
      },
...
    }
  }
}